requirements.txt
# python            3.13.1
# java              11.0.24
# npm               10.9.0
# node              22.12.0
# git               2.47.1

# Python version
python==3.13.1

# Python libraries
attrs==24.3.0
behave==1.2.6
certifi==2024.12.14
cffi==1.17.1
colorama==0.4.6
h11==0.14.0
idna==3.10
iniconfig==2.0.0
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==24.2
parse==1.20.2
parse_type==0.6.4
pluggy==1.5.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.4
pytweening==1.2.0
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0

# Additional libraries
webdriver_manager
cryptography
tenacity
openpyxl
slack-sdk
Appium-Python-Client